Fixed a race condition.
There's a time lag between

(A) FlowExecution.isComplete() starts returning true,
(B) Future from scheduleRun() signals completion, and
(C) WorkflowRun.finish() call happens to set the final result.

It used to be that the order in which this happens was (A) > (C) > (B)
because head updates and listener notifications happen before Future
signals.

With the change in the previous commit, now the order of the event is
(A) > (B) > (C), which breaks various tests that expect WorkflowRun
to have the final result when CpsFlowExecution completes.

This change fixes back the order.

Originally-Committed-As: 72fbbe238438beb4b1ee3062091c2059947bc15f