Run part details not exported via API
When part of a workflow run is executing on a node, the corresponding run details are not available via the API (xml/json)

Without this patch, the run part cannot be identified by examining the API call's result.

This poses a problem for an integration such as python-jenkins where its server.get_running_builds()
queries each node to build up an array of builds. As a mininum, 'number' and 'url' are needed.

The plugin exposes this data already via Jelly in the Executor view.